requirements.txt
asttokens==3.0.0
biopython==1.85
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.2
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
dill==0.4.0
executing==2.2.1
filelock==3.19.1
fonttools==4.60.1
globus-sdk==3.65.0
greenlet==3.2.4
idna==3.10
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
matplotlib==3.10.6
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.3.3
packaging==25.0
pandas==2.3.3
parsl==2025.9.29
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
requests==2.32.5
setproctitle==1.3.7
six==1.17.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.44
stack-data==0.6.3
tblib==3.1.0
tornado==6.5.2
traitlets==5.14.3
typeguard==4.4.4
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
